use goTo instead of goToDefinition
Since vscode has only one kind of 'goto' command, it makes sense
to use the one that does the "most sensible" thing.

From https://github.com/Valloric/YouCompleteMe#goto-commands
This command tries to perform the "most sensible" GoTo operation it can.
Currently, this means that it tries to look up the symbol under the cursor
and jumps to its definition if possible; if the definition is not accessible
from the current translation unit, jumps to the symbol's declaration.
For C/C++/Objective-C, it first tries to look up the current line for a header
and jump to it.