show diagnostic for errors in header file
If there are issues we come across in files other than the
one we're looking at, it's probably from an included header.
Since they may be the root source of errors in the file
we're looking at, instead of filtering them all out, let's
just pick the first one to display and hard-code it to
show up on the first line, since the language
server diagnostic interface doesn't appear to be able to
report errors in different files.