change activation events to be for languages c, cpp, and objective-c only
for people with ycmd.path set only in a workspace settings rather
than user settings, this change prevents the extension from complaining about
not being able to find ycmd when using vscode on an ad-hoc basis,
e.g. as the editor for git commit messages.  Also, it seems like the right
thing to do in general.