Enhance the logic for showing the hover popup (#45)
* Enhance the logic for showing the hover popup

The current getType call does not return useful information for
member functions.  getDoc returns the most useful info, for
things that have documentation, but nothing for things that
do not.  Next we try to find an exact completion for the
identifier and use its signature.  getType is appropriate
for the remaining cases.

This also adds the available documentation to the completion help.

* refine heuristic for extracting signature from getDoc

* Fix: typings